+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Changelog
 =========
 
 * Logic/module to link UCBShift for CS back-calculator (#10)
-* Documentation for installing third party applications
+* Documentation for installing UCBShift on top of SPyCi-PDB
 
 v0.0.9 (2022-08-10)
 ------------------------------------------------------------